requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
black==22.12.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coverage==7.2.3
cryptography==40.0.2
cssselect2==0.7.0
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.1.2
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-localflavor==3.1
django-parler==2.3
django-rosetta==0.9.8
django-stubs==1.16.0
django-stubs-ext==0.8.0
exceptiongroup==1.1.1
execnet==1.9.0
flake8==6.0.0
flake8-django==1.2
flake8-pydocstyle==0.2.2
flower==1.2.0
fonttools==4.37.4
freezegun==1.2.2
html5lib==1.1
humanize==4.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
isort==5.12.0
kombu==5.2.4
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
packaging==21.3
pathspec==0.11.1
Pillow==9.2.0
platformdirs==3.2.0
pluggy==1.0.0
polib==1.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.5
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pydyf==0.3.0
pyflakes==3.0.1
PyHamcrest==2.0.4
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyphen==0.13.0
pytest==7.3.1
pytest-celery==0.0.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-mock==3.10.0
pytest-xdist==3.2.1
python-dateutil==2.8.2
python-decouple==3.6
python-stdnum==1.18
pytz==2022.4
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.2
stripe==4.0.2
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
Twisted==22.10.0
txaio==23.1.1
types-psutil==5.9.5.11
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
urllib3==1.26.12
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
wrapt==1.14.1
zope.interface==6.0
zopfli==0.2.1
